test: only test relayChainStateOverrides on parachains
